Overhaul of assign roles dialog
Removed checkboxs and replaced with larger selectable pseudo-buttons. Added search functionality based off of the roles name.
If no roles are present on page, a message is shown to the user to indicate that. Updated functionality of listview-toolbar to
search on keystrokes instead of just on keyup.enter. Added a check for listview-toolbar to see if it was detached from a from in order to apply box-shadow. Added search functionality to role mappings under edit user.

squash this